Add new structure within tcn package
